requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
cloudinary==1.26.0
Django==3.2.7
django-cors-headers==3.9.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Faker==8.14.0
idna==3.2
jsonschema==4.1.2
Pillow==8.3.2
psycopg2-binary==2.9.1
Pygments==2.10.0
PyJWT==2.1.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
redgreenunittest==0.1.1
requests==2.26.0
rest-framework-smoke==0.4.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.7
gunicorn
whitenoise
